requirements.txt
# Core dependencies
numpy==1.24.3
pandas==1.5.0
matplotlib>=3.5.0
boto3>=1.24.0
apify-client>=1.0.0
requests>=2.28.0
pillow>=9.2.0
tqdm>=4.64.0
paramiko>=2.11.0
python-dotenv>=0.21.0
pytest>=7.0.0
pytest-cov>=4.0.0

# Machine learning dependencies
# torch>=1.13.1
# torchvision>=0.14.1
opencv-python-headless>=4.7.0
scikit-learn>=1.0.0
ultralytics==8.1.0
timm==0.8.13.dev0
lapx>=0.5.2